优化数据库Float类型判断错误

+++ smither-project-base/src/main/java/com/codesmither/project/base/constant/JavaLang.java
@@ -85,7 +85,7 @@ public class JavaLang extends ProgLang {
             case Types.OTHER:
                 return Object.class.getSimpleName();
             case Types.REAL:
-                return Double.class.getSimpleName();
+                return Float.class.getSimpleName();
             case Types.REF:
                 break;
             case Types.ROWID:
